+++ FixPointCS/Fixed32.cs
@@ -326,6 +326,9 @@ namespace FixPointCS
         [MethodImpl(FixedUtil.AggressiveInlining)]
         private static int Nlz(uint x)
         {
+        #if NET5_0_OR_GREATER
+            return System.Numerics.BitOperations.LeadingZeroCount(x);
+        #else
             int n = 0;
             if (x <= 0x0000FFFF) { n = n + 16; x = x << 16; }
             if (x <= 0x00FFFFFF) { n = n + 8; x = x << 8; }
@@ -334,6 +337,7 @@ namespace FixPointCS
             if (x <= 0x7FFFFFFF) { n = n + 1; }
             if (x == 0) return 32;
             return n;
+        #endif
         }
 #endif